requirements.txt
openai
python-dotenv
pandas
pyarrow
Pillow
tqdm